Fix clang tidy

+++ extensions/standard-processors/processors/DefragmentText.cpp
@@ -179,10 +179,10 @@ bool DefragmentText::splitFlowFileAtLastPattern(core::ProcessSession& session,
   }
   auto split_position = getSplitPosition(last_regex_match, pattern_location_);
   if (split_position != 0) {
-    split_before_last_pattern = session.clone(*original_flow_file, 0, split_position);
+    split_before_last_pattern = session.clone(*original_flow_file, 0, gsl::narrow<int64_t>(split_position));
   }
   if (split_position != original_flow_file->getSize()) {
-    split_after_last_pattern = session.clone(*original_flow_file, split_position, original_flow_file->getSize() - split_position);
+    split_after_last_pattern = session.clone(*original_flow_file, gsl::narrow<int64_t>(split_position), gsl::narrow<int64_t>(original_flow_file->getSize() - split_position));
   }
   updateAttributesForSplitFiles(*original_flow_file, split_before_last_pattern, split_after_last_pattern, split_position);
   return true;
